+++ src-tauri/src/main.rs
@@ -13,5 +13,5 @@ fn main() {
         std::env::set_var("WEBKIT_DISABLE_DMABUF_RENDERER", "1");
     }
 
-    paperling_lib::run()
+    mdtxt_lib::run()
 }